Delete Device By ID Endpoint Update
Temp route added for delete device by ID endpoint in security config file.

Delete controller updated to return deleted device entity object in success use-case. Returns 'null' in delete device error use-case.

+++ src/main/java/com/arnzen/home_api_backend/service/DeleteService.java
@@ -2,10 +2,13 @@ package com.arnzen.home_api_backend.service;
 
 import com.arnzen.home_api_backend.dao.DeviceDao;
 import com.arnzen.home_api_backend.dao.TemperatureDao;
+import com.arnzen.home_api_backend.model.DeviceEntity;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class DeleteService {
@@ -17,13 +20,20 @@ public class DeleteService {
     TemperatureDao temperatureDao;
 
     @Transactional
-    public boolean deleteDevice(int deviceId) throws EmptyResultDataAccessException {
+    public DeviceEntity deleteDevice(int deviceId) throws EmptyResultDataAccessException {
         try {
-            temperatureDao.deleteAllByDeviceEntityId(deviceId);
-            deviceDao.deleteById(deviceId);
-            return true;    // Device found in the database and removed.
+
+            Optional<DeviceEntity> deviceEntity = deviceDao.findById(deviceId);
+
+            if (deviceEntity.isPresent()) {
+                temperatureDao.deleteAllByDeviceEntityId(deviceId);
+                deviceDao.deleteById(deviceId);
+                return deviceEntity.get();    // Device found in the database and removed.
+            } else {
+                return null; // Device not found in the database. No error thrown.
+            }
         } catch (EmptyResultDataAccessException exception) {
-            return false;   // Device not found in the database and error thrown.
+            return null;   // Device not found in the database and error thrown.
         }
     }
 }
